requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.11.0
Django==4.2.1
django-dotenv==1.4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.12.0
fonttools==4.39.4
frozenlist==1.3.3
idna==3.4
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numpy==1.24.3
openai==0.27.7
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
psutil==5.9.5
PyJWT==2.7.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
scipy==1.10.1
seaborn==0.12.2
sentry-sdk==1.23.1
six==1.16.0
sqlparse==0.4.4
sympy==1.12
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
typing_extensions==4.6.0
tzdata==2023.3
ultralytics==8.0.106
urllib3==1.26.15
yarl==1.9.2
